refactor(shared): refactor code #181
